refactor_config.go
In this update, major changes have been incorporated to the `config to make it more maintainable and easily understandable. go` script. Some of the changes made are: The function Check Map Invariant has been refactored to decrease the cognitive complexity of the code since it was initially high and made code maintenance difficult. It has allowed for the function to be more easily understood and maintained in a more modular manor by dividing it into smaller sub-functions. This refactoring involved the creation of helper functions that would be used in validating the configuration object storage, handling errors and updating elements. Moreover, we have improved the error management system concerning creation or modification and deletion of configuration data objects. These features provide for enhanced error checking and handling mechanisms hence decreasing the chances of errors when undertaking configuration tasks. The script has been brought in to conform to good coding standards, which improves the readability, test ability and extensibility of the code for future enhancement and support. In this way, these improvements enhance the code quality and make the code more maintainable and more in line with the standard software development practices.